Manage match not played, having a winner
Previously, in the `PlayerMatch` table, when the match was not played
(results in columns `teamXsetY` are 0), we did not expect to have a winner
(`winner` column == 0).
In this case, an error was raised (in the `Match` model).

However, it seems this usecase can really exist:
i.e. in an `elimination` draw, a player can be exempted for the 1st round.

This commit does not raise an error in such a use case.