Apply "smart" sorting on clubs/players records
Previously, clubs/players records were always sorterd by points (descending),
which make sense when the `show_ranking` URL parameter is set.
But when not set, it is more relevant to sort records by their name, which is
what this commit does.

Note:
The change done in the `get_ranks` function (parameter `sort_by` added) allows
to easily sort records by any other attribute of a club/player.
For instance, for players, we could imagine sorting by `firstname` instead of
`name` (which is currently the case).