[REF][WIP] Factorize html tables
Within the entire repository multiple tables with the same header properties
are used.
This commit implements a table template which can be included everywhere a
table is used.

Note:
The way the template has been created is a bit dirty as it contains only the
opening `table` tag and nothing else.
This is because it can then be used with jinja2 `include` instruction.
However, we could consider a better solution (as using the `extends`
instruction). But as of now, using this solution has been been unsuccessful.

[TODO] Currently, only one tables uses this template. Need to factorize for
remaining tables.